Add interface memory management and improve string handling
- Introduced a single block allocation for interface names to optimize memory usage.
- Updated `free_ifaces` to free the allocated block instead of individual interface strings.
- Added `trim_token` function to clean up whitespace from tokens during interface parsing.
- Enhanced `set_interfaces_from_string` to utilize the new allocation and improved token handling.
- Adjusted various functions to ensure proper handling of sizes and types, preventing potential overflows and improving robustness.

+++ sendfile.c
@@ -29,7 +29,10 @@
 #include "utils.h"
 #include "log.h"
 
+/* Max bytes per sendfile() call (2^31-1): keeps count in 32-bit range, avoids huge
+ * single kernel transfers. Only used when HAVE_SYS_SENDFILE. */
 #define MAX_BUFFER_SIZE 2147483647
+/* Fallback read/write buffer size (64 KiB): one allocation, decent throughput. */
 #define MIN_BUFFER_SIZE 65536
 
 #if defined(__linux__)
@@ -79,9 +82,9 @@ static inline int sys_sendfile(int sock, int sendfd, off_t *offset, off_t len)
 
 
 
-void send_file(int socketfd, int sendfd, off_t offset, size_t end_offset)
+void send_file(int socketfd, int sendfd, off_t offset, off_t end_offset)
 {
-    size_t send_size;
+    off_t send_size;
     off_t ret;
 
 #if defined(HAVE_SYS_SENDFILE)
@@ -95,9 +98,9 @@ void send_file(int socketfd, int sendfd, off_t offset, size_t end_offset)
             if (send_size > MAX_BUFFER_SIZE)
                 send_size = MAX_BUFFER_SIZE;
 
-            PRINT_LOG(E_DEBUG, "sendfile range %jd to %zu\n", (intmax_t)offset,
-                      send_size);
-            ret = sys_sendfile(socketfd, sendfd, &offset, send_size);
+            PRINT_LOG(E_DEBUG, "sendfile range %jd to %jd\n", (intmax_t)offset,
+                      (intmax_t)send_size);
+            ret = sys_sendfile(socketfd, sendfd, &offset, (size_t)send_size);
             if (ret == -1)
             {
                 // a broken pipe isn't really an error
@@ -126,11 +129,11 @@ fallback:
     while (offset <= end_offset)
     {
         send_size = end_offset - offset + 1;
-        if (send_size > MIN_BUFFER_SIZE)
+        if (send_size > (off_t)MIN_BUFFER_SIZE)
             send_size = MIN_BUFFER_SIZE;
 
         lseek(sendfd, offset, SEEK_SET);
-        ret = read(sendfd, buf, send_size);
+        ret = read(sendfd, buf, (size_t)send_size);
         if (ret == -1)
         {
             PRINT_LOG(E_DEBUG, "read error :: error no. %d\n", errno);
